Patch from ryan.freckleton@gmail.com: Improved spec formatting when using docstrings
The inspect module is now used on docstrings of objects instead of __doc__.
This dedents the docstring so that multiline docstrings are displayed with
nicer formatting when using --with-spec.

in_color was also modified to colorize all the multiple lines of docstring
based test names. This allows colors to be correctly displayed when using the
less -R command.

Another nicety would be to add the test status to the first line of the test
description, currently it is appeneded to the end.