Make stopwatch work with test methods.
Prior to this commit, test methods are always run, even if they are
slower than the "--faster-than" threshold.

The problem was, stopTest put the qualified name of the test class in
the dontrun dict, instead of the qualified name of the *method*.

To fix this, we use the id() method of test to get the qualified name,
just as xunit, a built-in plugin to nose, does.